requirements.txt
fastapi
pydantic
schemas
psycopg2-binary